chore: Update dependencies and package configurations
The order of the examples matters. examples/cli is alphabetically earlier than examples/web, so the cli version of react will be installed in the root node_modules, and web will be local. But examples/web uses Vite, and that will look at the root regardless of the local version. So copy the version of react et al in examples/web and put it in the root packages.json. you will end up with multiple copies, but that is ok. Now I use ink in examples/cli. It gets installed at root level so looks there and gets web's version of react. But if you put an old version of ink at root level, then the real version gets installed local to package cli and will use the local version of react there and it all works.